feat: update Patient model and controller for improved patient profile handling

+++ patient-service/src/main/java/com/patientmanagement/patientservice/repository/PatientRepository.java
@@ -1,15 +1,11 @@
 package com.patientmanagement.patientservice.repository;
 
-import com.patientmanagement.patientservice.dto.PatientResponseDTO;
 import com.patientmanagement.patientservice.model.Patient;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
-public interface PatientRepository extends JpaRepository<Patient, Integer> {
+public interface PatientRepository extends JpaRepository<Patient, Long> {
 
 
     boolean existsByUsername(String username);
